requirements.txt
google-search-results>=2.4.2
psycopg2-binary>=2.9.9
